+++ test/support/no_native_registry.ex
@@ -1,4 +1,4 @@
 defmodule ExCellerate.Test.NoNativeRegistry do
   @moduledoc false
-  use ExCellerate.Registry, native_compilation: false
+  use ExCellerate.Registry, compilation: ExCellerate.Compilation.Interpreted
 end